initial progress on grab bag of tickets
